Fix DefaultSignatures context for GHC 8.0.2
GHC 8.0.2 fixes some behavior around default signature contexts and as a
result is stricter about the sort of constraints that can be applied to
default signatures [0]. This causes the default dataFetch instances to
fail to unify with messages like:

        Expected type: StateT s m a
          Actual type: StateT s (StateT s m) a

The consensus on that ticket is that the 8.0.2 behavior is correct, and
Simon PJ suggests that default signatures should only ever change the
context of methods, never the structure of the type itself.

This commit fixes the build on 8.0.2 by factoring out the `t m a` into
an equational constraint.

[0] https://ghc.haskell.org/trac/ghc/ticket/12784